Fix checkstyle and break willingly
