+++ src/me/corriekay/pokegoutil/utils/pokemon/PokemonPerformanceCache.java
@@ -101,68 +101,4 @@ public final class PokemonPerformanceCache {
     }
 
 
-    /**
-     * Holds the Duel Ability, Gym Offense and Gym Defense Perfomances for given Pokémon.
-     */
-    public static final class PokemonPerformanceStats {
-        public final PokemonId pokemonId;
-        public final PokemonPerformance duelAbility;
-        public final PokemonPerformance gymOffense;
-        public final PokemonPerformance gymDefense;
-
-        /**
-         * Creates an instance of this performance stats object.
-         * This is just an internal data class.
-         *
-         * @param pokemonId   The Pokémon ID.
-         * @param duelAbility The Duel Ability Performance.
-         * @param gymOffense  The Gym Offense Performance.
-         * @param gymDefense  The Gym Defense Performance.
-         */
-        private PokemonPerformanceStats(final PokemonId pokemonId,
-                                        final PokemonPerformance duelAbility,
-                                        final PokemonPerformance gymOffense,
-                                        final PokemonPerformance gymDefense) {
-            this.pokemonId = pokemonId;
-            this.duelAbility = duelAbility;
-            this.gymOffense = gymOffense;
-            this.gymDefense = gymDefense;
-        }
-    }
-
-    /**
-     * A class to save a specific performance of a given Pokémon with given moveset.
-     */
-    public static final class PokemonPerformance {
-        public static final PokemonPerformance DEFAULT = new PokemonPerformance(null, 0, null, null);
-
-        public final PokemonId pokemonId;
-        public final double value;
-        public final PokemonMove move1;
-        public final PokemonMove move2;
-
-        /**
-         * Creates an instance of this performance stats object.
-         * This is just an internal data class.
-         *
-         * @param pokemonId The Pokémon ID.
-         * @param value     The value for this performance
-         * @param move1     The Primary Move.
-         * @param move2     The Secondary Move.
-         */
-        private PokemonPerformance(final PokemonId pokemonId, final double value, final PokemonMove move1, final PokemonMove move2) {
-            this.pokemonId = pokemonId;
-            this.value = value;
-            this.move1 = move1;
-            this.move2 = move2;
-        }
-
-        @Override
-        public String toString() {
-            return PokemonUtils.getLocalPokeName(pokemonId.getNumber()) + " with "
-                + PokemonUtils.formatMove(move1) + " and "
-                + PokemonUtils.formatMove(move2) + " has "
-                + value;
-        }
-    }
 }
