+++ src/me/corriekay/pokegoutil/utils/windows/PokeColumn.java
@@ -278,8 +278,10 @@ public enum PokeColumn {
      * @return The column.
      */
     public static PokeColumn getForId(final int id) {
+        // TODO: We break checkstyle willingly, to see that this check is working
+        final int data = id;
         for (final PokeColumn column : PokeColumn.values()) {
-            if (column.id == id) {
+            if (column.id == data) {
                 return column;
             }
         }
